placed header and image holder
