+++ js/attractionTypesToDom.js
@@ -16,15 +16,22 @@ function putAttractionTypesToDom (result) {
     // let cardDeck = document.createElement("div");
     // cardDeck.className = "card-deck justify-content row";
     ///////putting in the loop///////
+    let header = document.createElement('h1');
+    header.className = "col-md-12"
+    header.innerText = "Attraction Varieties";
+    fragment.appendChild(header);
     let count = 0;
     result.forEach(function (type) {
         let card = document.createElement("div");
         card.className = "card col-sm-4";
-
+        let createImgHolder = document.createElement("div");
+        createImgHolder.className = "image-holder";
         let img = document.createElement('img');
         img.className = "card-img-top";
         img.src = `${images[count]}`;
-        card.appendChild(img);
+        createImgHolder.appendChild(img);
+        card.appendChild(createImgHolder);
+
         let cardBody = document.createElement("div");
         cardBody.className = "card-body";
         cardBody.innerHTML = `<h5 class="card-title">${capitalLetter(type.name)}</h5>`;
